refactor: add settings sidebar and revamp shop and address settings
