+++ imports/plugins/core/payments/client/PaymentSettings.js
@@ -7,7 +7,6 @@ import Button from "@material-ui/core/Button";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import CardHeader from "@material-ui/core/CardHeader";
-import { Blocks } from "@reactioncommerce/reaction-components";
 import { i18next } from "/client/api";
 import useCurrentShopId from "/imports/client/ui/hooks/useCurrentShopId.js";
 
@@ -105,7 +104,6 @@ export default function PaymentSettings(props) {
           ))}
         </CardContent>
       </Card>
-      <Blocks region="PaymentSettings" blockProps={props} />
     </div>
   );
 }
