+++ imports/plugins/core/payments/client/PaymentSettingsRegion.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Blocks } from "@reactioncommerce/reaction-components";
+
+/**
+ * @summary Renders payment settings page
+ * @param {Object} props Component props
+ * @return {React.Node} React node
+ */
+export default function PaymentSettingsRegion(props) {
+  return (
+    <Blocks region="PaymentSettings" blockProps={props} />
+  );
+}
