+++ imports/plugins/core/payments/client/index.js
@@ -1,16 +1,20 @@
-import React from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCreditCard } from "@fortawesome/free-solid-svg-icons";
 import { registerOperatorRoute } from "/imports/client/ui";
-import PaymentSettings from "./PaymentSettings.js";
+import { registerBlock } from "@reactioncommerce/reaction-components";
+import PaymentSettingsRegion from "./PaymentSettingsRegion";
+import PaymentSettings from "./PaymentSettings";
 
 registerOperatorRoute({
   isNavigationLink: true,
   isSetting: true,
-  mainComponent: PaymentSettings,
+  mainComponent: PaymentSettingsRegion,
   priority: 20,
-  path: "/payment",
-  // eslint-disable-next-line react/display-name
-  SidebarIconComponent: (props) => <FontAwesomeIcon icon={faCreditCard} {...props} />,
+  path: "/settings/payment",
   sidebarI18nLabel: "admin.settings.paymentSettingsLabel"
+});
+
+registerBlock({
+  region: "PaymentSettings",
+  name: "PaymentSettingsDefault",
+  component: PaymentSettings,
+  priority: 1
 });
